feat: rename settings tab to two-factor auth, split webauthn component
